requirements.txt
Flask==2.0.1
requests==2.31.0
debugpy
google-generativeai==0.3.2
pytube==15.0.0
python-dotenv==1.0.1
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
gunicorn==20.1.0
youtube-transcript-api==0.6.1
playwright==1.42.0
grpcio==1.60.0
yt-dlp==2023.12.30
sumy==0.11.0
transformers==4.37.2
konlpy==0.6.0
google-api-python-client==2.108.0 